One class per file Fixes #13
